Major scene refactor
While trying to change the way scenes are loaded I created the new example folder which should be able to hold
any future examples I need to create, the Scene class for new is a virtual class that will be used in the
windows class. A scene can have whatever is need to be displayed in the engine, it is not final but its a more
manageable solution then a simple main file.