Created a new concept the Renderable2D
The renderable and renderer should both work together to manage all the drawable types that the engine is going
to have in the future, the sprite is now a renderable which will be controlled by the renderer with this new
concept I need to go back to the sprite and clean more the class. I also moved the logger to a new utils folder
which should that more files as I create more supporting tools.