Final configuration settings to both engine and game
Removed the post build scripts to copy the dll to the build folder of the engine, they now
work on the game project, removed the auto lib linking comment for opengl32.lib this is now
used in the linker additional libraries. Added the build path of the engine to the game
because of the build lib file.